+++ backend/core/src/seeds/listings/listing-triton-seed.ts
@@ -820,3 +820,57 @@ export class ListingTritonSeed extends ListingDefaultSeed {
     return await this.listingRepository.save(listingCreateDto)
   }
 }
+
+export class ListingTritonSeedDetroit extends ListingDefaultSeed {
+  async seed() {
+    const unitTypeOneBdrm = await this.unitTypeRepository.findOneOrFail({ name: "oneBdrm" })
+    const unitTypeTwoBdrm = await this.unitTypeRepository.findOneOrFail({ name: "twoBdrm" })
+
+    const detroitJurisdiction = await this.jurisdictionRepository.findOneOrFail({
+      name: CountyCode.detroit,
+    })
+    const amiChart = await this.amiChartRepository.save({
+      ...tritonAmiChart,
+      name: `${tritonAmiChart.name}Detroit`,
+      jurisdiction: detroitJurisdiction,
+    })
+
+    const property = await this.propertyRepository.findOneOrFail({
+      developer: "Thompson Dorfman, LLC",
+      neighborhood: "Foster City",
+      smokingPolicy: "Non-Smoking",
+    })
+
+    const unitsToBeCreated: Array<Omit<UnitCreateDto, keyof BaseEntity>> = tritonUnits.map(
+      (unit) => {
+        return {
+          ...unit,
+          property: {
+            id: property.id,
+          },
+          amiChart,
+        }
+      }
+    )
+
+    unitsToBeCreated[0].unitType = unitTypeTwoBdrm
+    unitsToBeCreated[1].unitType = unitTypeOneBdrm
+    unitsToBeCreated[2].unitType = unitTypeOneBdrm
+    unitsToBeCreated[3].unitType = unitTypeOneBdrm
+    unitsToBeCreated[4].unitType = unitTypeOneBdrm
+
+    await this.unitsRepository.save(unitsToBeCreated)
+
+    const listingCreateDto: Omit<
+      DeepPartial<Listing>,
+      keyof BaseEntity | "urlSlug" | "showWaitlist"
+    > = {
+      ...tritonListing,
+      property: property,
+      assets: getDefaultAssets(),
+      events: [],
+    }
+
+    return await this.listingRepository.save(listingCreateDto)
+  }
+}